requirements.txt
discord.py==1.5.1
python-dotenv==0.12.0
